chore: refactor chat out of app component
